+++ volmdlr/core.py
@@ -85,7 +85,7 @@ def step_ids_to_str(ids):
     :return: A string containing all the IDs
     :rtype: str
     """
-    return ','.join(['#{}'.format(i) for i in ids])
+    return ','.join([f"#{i}" for i in ids])
 
 
 class CompositePrimitive(dc.PhysicalObject):
@@ -105,7 +105,12 @@ class CompositePrimitive(dc.PhysicalObject):
 
         dc.PhysicalObject.__init__(self, name=name)
 
+    def to_dict(self, *args, **kwargs):
+        """Avoids storing points in memo that makes serialization slow."""
+        return dc.PhysicalObject.to_dict(self, use_pointers=False)
+
     def primitive_to_index(self, primitive):
+        """Constructs a dictionary associating primitive to its index."""
         if not self._utd_primitives_to_index:
             self._primitives_to_index = {p: ip for ip, p in enumerate(self.primitives)}
             self._utd_primitives_to_index = True
@@ -280,8 +285,7 @@ class Primitive3D(dc.PhysicalObject):
 
     def triangulation(self):
         raise NotImplementedError(
-            'triangulation method should be implemented on class {}'.format(
-                self.__class__.__name__))
+            f"triangulation method should be implemented on class {self.__class__.__name__}")
 
     def babylon_meshes(self):
         mesh = self.triangulation()
@@ -574,7 +578,7 @@ class BoundingBox(dc.DessiaObject):
                            min(self.zmin, other_bbox.zmin),
                            max(self.zmax, other_bbox.zmax))
 
-    def to_dict(self, use_pointers: bool = True, memo=None, path: str = '#') -> dict:
+    def to_dict(self, *args, **kwargs) -> dict:
         """
         Converts the bounding box to a dictionary representation.
 
@@ -1234,7 +1238,7 @@ class VolumeModel(dc.PhysicalObject):
         """
         Gets the lines that define a VolumeModel geometry in a .geo file.
 
-        :return: A list of lines that describe the geomery
+        :return: A list of lines that describe the geometry
         :rtype: List[str]
 
         """
@@ -1704,7 +1708,6 @@ class MovingVolumeModel(VolumeModel):
                 primitives_to_meshes.append(i_prim)
 
         bbox = self._bounding_box()
-        center = bbox.center
         max_length = max([bbox.xmax - bbox.xmin,
                           bbox.ymax - bbox.ymin,
                           bbox.zmax - bbox.zmin])
@@ -1728,6 +1731,6 @@ class MovingVolumeModel(VolumeModel):
 
         babylon_data = {'meshes': meshes,
                         'max_length': max_length,
-                        'center': list(center),
+                        'center': list(bbox.center),
                         'steps': steps}
         return babylon_data